requirements.txt
tensorflow==1.15
scipy==1.1.0
scikit-learn
opencv-python
h5py
matplotlib
Pillow
requests
psutil
playsound==1.2.2
wxPython